% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{annotate}
\alias{annotate}
\title{Add the annotations to a Seurat object}
\usage{
annotate(seu, annotations, to_label, annotation_name = "CellType")
}
\description{
This function takes in a column from a data frame or vector of cell type
annotations. For example annotations = df$CellTypes or annotations = c("Neurons",
"astrocytes","radial glia").The data slot to annotate must be indicated.
To label the data slot with the consensus annotation set annotation_name.
Default annotation_name = "CellType".
Example: seu <- annotate(seu, annotations = df$CellTypes, to_label = seu$clusters, annotation_name = "CellTypes")
}
