% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{annotate_df}
\alias{annotate_df}
\title{This function takes in a list of annotation data frames
The outputs of get_annotations or a manually created data frame
The colnames must be "Cluster" and a different column name for the prediction method.}
\usage{
annotate_df(ann.list)
}
\description{
This function takes in a list of annotation data frames
The outputs of get_annotations or a manually created data frame
The colnames must be "Cluster" and a different column name for the prediction method.
}
