% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{clust_stability}
\alias{clust_stability}
\title{Test the stability or reproducibility of clusters across kn or resolutions
For Louvain clustering in a seurat object test calculate the RAND Index. For n iterations
each cluser identity for each cell is compared with the other iterations of clustering.
Select one pararmeter to vary resolutions = c(0.1,0.3,0.5,0.8) or kn = c(20,40,60,80).
The function also calculates the mean and standard deviation of the number of clusters
for each resolution or kn. The function returns an outputs and plot of the results.}
\usage{
clust_stability(input, resolutions, kn, pcdim = 1:10, n = 100, save_to = NULL)
}
\description{
Test the stability or reproducibility of clusters across kn or resolutions
For Louvain clustering in a seurat object test calculate the RAND Index. For n iterations
each cluser identity for each cell is compared with the other iterations of clustering.
Select one pararmeter to vary resolutions = c(0.1,0.3,0.5,0.8) or kn = c(20,40,60,80).
The function also calculates the mean and standard deviation of the number of clusters
for each resolution or kn. The function returns an outputs and plot of the results.
}
