% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{explore_param}
\alias{explore_param}
\title{Explore different clustering methods and parameters in single cell data
Takes in a Seurat object. Choose a clustering method or vector of clustering methods:
cluster_method = c("louvain","flowsom","phenograph"). The flowsom method requires the
original data frame of the Flow Cytometry expression data. For flows k is the number of
final clusters. To explore different cluster numbers flow_k = c(1,5,10,15). For louvain and
phenograph (phenograph uses Louvain clustering with a Jaccard index) the kn value is the
starting value for the snn nearest neighbour network detection: pheno_lou_kn = c(20, 60, 200).
For louvain network detection run with the Seurat function a vector of resolutions can be
input: lou_resolution = (0,0.2,0.5,1.2). If run.plot = TRUE the results will be shown in
using the clustree function.  Putting a starting resolution of 0 will make the visualization
better. Setting run.plot = TRUE will also produce UMAPs and heatmaps for each cluster
condition. If the option run.stats = TRUE internal statistics will be calculated. Data tables will be
produced in a list. If you wish to save the plots generated and the list of statistic set
save_to = filepath/tosaveto/ . To run intrinsic statistics. Silhouette score ranges from -1 to 1.
A value near -1 indicates a poor quality of the clusters, a value near 0 indicates the neutrality and
insignificance of the clusters, and a value near 1 indicates a good quality of the clusters
(Rousseeuw, 1987). Calinski-Harabasz index does not have a range, but a higher index indicates a
better cluster. The number of clusters with the highest Calinski-Harabasz index is the optimal number
of clusters (Calinski & Harabasz, 1974). Davies-Bouldin index has no upper bound but the minimum index
is 0. A lower index indicates better clusters (Davies & Bouldin, 1979).}
\usage{
explore_param(
  input,
  cluster_method,
  df_input,
  flow_k = NULL,
  pheno_lou_kn = NULL,
  lou_resolution = NULL,
  pcdim = 1:10,
  run.plot = TRUE,
  run.stats = TRUE,
  save_to = NULL
)
}
\description{
Explore different clustering methods and parameters in single cell data
Takes in a Seurat object. Choose a clustering method or vector of clustering methods:
cluster_method = c("louvain","flowsom","phenograph"). The flowsom method requires the
original data frame of the Flow Cytometry expression data. For flows k is the number of
final clusters. To explore different cluster numbers flow_k = c(1,5,10,15). For louvain and
phenograph (phenograph uses Louvain clustering with a Jaccard index) the kn value is the
starting value for the snn nearest neighbour network detection: pheno_lou_kn = c(20, 60, 200).
For louvain network detection run with the Seurat function a vector of resolutions can be
input: lou_resolution = (0,0.2,0.5,1.2). If run.plot = TRUE the results will be shown in
using the clustree function.  Putting a starting resolution of 0 will make the visualization
better. Setting run.plot = TRUE will also produce UMAPs and heatmaps for each cluster
condition. If the option run.stats = TRUE internal statistics will be calculated. Data tables will be
produced in a list. If you wish to save the plots generated and the list of statistic set
save_to = filepath/tosaveto/ . To run intrinsic statistics. Silhouette score ranges from -1 to 1.
A value near -1 indicates a poor quality of the clusters, a value near 0 indicates the neutrality and
insignificance of the clusters, and a value near 1 indicates a good quality of the clusters
(Rousseeuw, 1987). Calinski-Harabasz index does not have a range, but a higher index indicates a
better cluster. The number of clusters with the highest Calinski-Harabasz index is the optimal number
of clusters (Calinski & Harabasz, 1974). Davies-Bouldin index has no upper bound but the minimum index
is 0. A lower index indicates better clusters (Davies & Bouldin, 1979).
}
\examples{
\dontrun{
explore_param(seuratObj, cluster_method = "louvain",df_input = flowset_dataframe, flow_k = NULL,
pheno_lou_kn = 100, lou_resolution = c(0,0.5,1), pcdim = 1:10)
}
}
