% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R, R/update_CelltypeR_functions.R
\name{flowset_to_csv}
\alias{flowset_to_csv}
\title{Create a dataframe from a flowset object and option to save csv
Takes in a flowset object and creates a dataframe. By default no output path is added and a data frame is returned.
The name of each sample will be added as the column "Sample"
into the final dataframe. To save the dataframe directly as a csv set save.csv = filepath/tosave/
The filename will be the name of the flowset object input into the function.
flowset_to_csv(flowset)
flowset_to_csv(flowset, output_path = "path/to/location/", save.csv = TRUE)}
\usage{
flowset_to_csv(
  flowset,
  output_path = NULL,
  save.csv = FALSE,
  select_measure = "Area"
)

flowset_to_csv(
  flowset,
  output_path = NULL,
  save.csv = FALSE,
  select_measure = "Area"
)
}
\arguments{
\item{flowset}{A flowSet object}

\item{output_path}{Path to save CSV (if save.csv = TRUE)}

\item{save.csv}{Logical, whether to save CSV}

\item{select_measure}{One of c("all", "Area", "Height", "Width"). Default "Area"}
}
\description{
Create a dataframe from a flowset object and option to save csv
Takes in a flowset object and creates a dataframe. By default no output path is added and a data frame is returned.
The name of each sample will be added as the column "Sample"
into the final dataframe. To save the dataframe directly as a csv set save.csv = filepath/tosave/
The filename will be the name of the flowset object input into the function.
flowset_to_csv(flowset)
flowset_to_csv(flowset, output_path = "path/to/location/", save.csv = TRUE)

Convert flowSet to a single dataframe with optional selection of measurement type
}
\examples{
\dontrun{
flowset_to_csv(flowset)
flowset_to_csv(flowset, output_path = "path/to/location/", save.csv = TRUE)
}
}
