% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{harmonize}
\alias{harmonize}
\title{Transform and align samples in flowset object
This function to takes in the flowset object created from a fsc files and transformed the data
If 'retro' (default) is selected the samples will be biexponentially transformed, aligned and then reverse transformed
If 'biexp' is selected the samples will only be biexponentially transformed
If 'align' is selected the samples will be transformed and aligned but not reverse transformed
The two_peaks and one_peak argument refer to the biexp transformed data and are the indexes of which measures have
one or two peaks. At least one measurement of two peaks must be included. You will receive an error if there is only one peak in a measurement defined as having two peaks.
retro' which will perform biexp transform, align peaks and reverse transform
biexp' will preform a biexp transform
align' will perform a biexp transform and align peaks}
\usage{
harmonize(
  flowset,
  processing = "retro",
  two_peaks = c(9:length(colnames(transformed_flowset))),
  one_peak = c(1:8),
  threshold = 0.01
)
}
\description{
Transform and align samples in flowset object
This function to takes in the flowset object created from a fsc files and transformed the data
If 'retro' (default) is selected the samples will be biexponentially transformed, aligned and then reverse transformed
If 'biexp' is selected the samples will only be biexponentially transformed
If 'align' is selected the samples will be transformed and aligned but not reverse transformed
The two_peaks and one_peak argument refer to the biexp transformed data and are the indexes of which measures have
one or two peaks. At least one measurement of two peaks must be included. You will receive an error if there is only one peak in a measurement defined as having two peaks.
retro' which will perform biexp transform, align peaks and reverse transform
biexp' will preform a biexp transform
align' will perform a biexp transform and align peaks
}
\examples{
\dontrun{
harmonize(flowset, processing = 'retro', two_peaks = c(10:20), one_peak = c(1:9), theshold = 0.01)
}
}
