% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{make_seu}
\alias{make_seu}
\title{Creates a seurat object from the Flow Cytometry expression data frame.
Takes in a dataframe created from Flow Cytometry data using the fsc_to_fs and fsc_to_csv
functions. This function creates a Seurat object from the Flow Cytometry expression data
in a data frame format where cells are rows and Markers are columns.  A column indicating
the starting fsc file "Sample" is required and added as meta data into the Seurat object.
This now accounts for negative values that become NaN}
\usage{
make_seu(df, AB_vector)
}
\description{
Creates a seurat object from the Flow Cytometry expression data frame.
Takes in a dataframe created from Flow Cytometry data using the fsc_to_fs and fsc_to_csv
functions. This function creates a Seurat object from the Flow Cytometry expression data
in a data frame format where cells are rows and Markers are columns.  A column indicating
the starting fsc file "Sample" is required and added as meta data into the Seurat object.
This now accounts for negative values that become NaN
}
\examples{
\dontrun{
make_seu(df = flow_dataframe, AB_vector = markers_names)
}
}
