% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{permutation_test_multi}
\alias{permutation_test_multi}
\title{Permatation test multiple conditions using ANOVA
Input single cell data in the from of sc_utils_obj
The object must be made from the function sc_utils function in the library scProprotionTest
The meta data from cluster annotations and the meta data slot containing
the dependent variables to compare are entered. Three or more conditions must be present.
A loop running the anova permutation test from the library permuco is run for each cell type.
Example: permutation_avo <- permutation_test_multi(sc_utils_obj = sc_utils_obj,
cluster_identity = "Celltypes",sample_identity = "IPSC")}
\usage{
permutation_test_multi(
  sc_utils_obj,
  cluster_identity = "Celltypes",
  sample_identity = "Batch",
  n_permutations = 3000
)
}
\description{
Permatation test multiple conditions using ANOVA
Input single cell data in the from of sc_utils_obj
The object must be made from the function sc_utils function in the library scProprotionTest
The meta data from cluster annotations and the meta data slot containing
the dependent variables to compare are entered. Three or more conditions must be present.
A loop running the anova permutation test from the library permuco is run for each cell type.
Example: permutation_avo <- permutation_test_multi(sc_utils_obj = sc_utils_obj,
cluster_identity = "Celltypes",sample_identity = "IPSC")
}
