% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{plot_corr}
\alias{plot_corr}
\title{Plot correlation assignment model (CAM) results}
\usage{
plot_corr(df, threshold = 0, min_cells = 100)
}
\description{
The function requires the data frame output from the find_correlation function
Filter out cell type annotation with fewer than 'min_cells' for the plot
The threshold with draw a line on the plot.  This can be set to the R theshold used in the find_correlation function.
}
\examples{
\dontrun{
plot_corr(cor_df, threshold = 0.3, min_cells = 300)
}
}
