% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{plot_lab_clust}
\alias{plot_lab_clust}
\title{Creates a UMAP and barchart to visualize cell type predictions}
\usage{
plot_lab_clust(
  seu,
  seu.cluster,
  seu.labels,
  filter_out = c("unknown", "Unknown", "Mixed")
)
}
\description{
Takes a Seurat object with predicted cell type meta data slot(s) and cluster. Creates
a UMAP with the predictions labelled and a bar chart with the proportion of predicted
cell types in each cluster. There is an option to ignore the selected cell types using
filter_out. The followin inputs are required:
seu = Seurat_object
seu.cluster = seu$RNA_snn_res.0.5 (the metadata slot with the clusters you want to annotate)
seu.labels = seu$seu.pred (the metadata slot with the prediciton or labels to visualize)
filter_out = c("unknown") enter the annotation to filter out or set
filter_out = "none" to leave all annotations.
}
