% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{plot_randindex}
\alias{plot_randindex}
\title{Plot cluster stability results from output data}
\usage{
plot_randindex(rdf, cp = c("orange", "violet"), view = c(0, 1))
}
\description{
Requires the rdf object output from the function clust_stability. The colours and
viewable area of the plot can be altered. cp takes a vector of two colours. view takes
vector of for the x axis of two coordinates.
}
