% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{plotdensity_flowset}
\alias{plotdensity_flowset}
\title{Create a density plot from a flowset object
Takes in a flowset object. Plot density plots for each channel in multichannel fsc
Flow Cytometry data. Creates stacked plots where each row is a sample in the
set object.}
\usage{
plotdensity_flowset(
  flowset,
  sample_include = "all",
  select_channels = "all",
  scale_factor = 50,
  max_name_length = 20,
  facet_text_size = 6
)
}
\description{
Create a density plot from a flowset object
Takes in a flowset object. Plot density plots for each channel in multichannel fsc
Flow Cytometry data. Creates stacked plots where each row is a sample in the
set object.
}
