% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{plotmean}
\alias{plotmean}
\title{Plot mean values by group in dotplot or heatmap
This function takes a Seurat object and creates a heatmap or dotplot.
This function takes a Seurat object and a list of variables to plot. One
var_names is a vector with the levels of the seurat groups to be plotted
group is the cell types or x axis grouping variable
cluster_order is a character vector that can be used to reorder cell type annotations
marker_order is the order to plot the markers}
\usage{
plotmean(
  plot_type = "heatmap",
  seu,
  group,
  markers,
  var_names,
  input_assay = "RNA",
  xlab = "Cell Types",
  ylab = "Markers",
  cluster_order = NULL,
  marker_order = NULL,
  low_colour = "grey",
  mid_colour = "white",
  high_colour = "red"
)
}
\description{
Plot mean values by group in dotplot or heatmap
This function takes a Seurat object and creates a heatmap or dotplot.
This function takes a Seurat object and a list of variables to plot. One
var_names is a vector with the levels of the seurat groups to be plotted
group is the cell types or x axis grouping variable
cluster_order is a character vector that can be used to reorder cell type annotations
marker_order is the order to plot the markers
}
