% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{RFM_predict}
\alias{RFM_predict}
\title{Predict cell types in a Seurat object with a trained Random Forest Model}
\usage{
RFM_predict(seu, rf, markers)
}
\description{
Predict cell types in a Seurat object using a trained Random Forest Model.
A trained model is required. Returns the Seurat object with the predictions
added as the metadata slot 'rfm.labels'.
}
