% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{seurat_predict}
\alias{seurat_predict}
\title{Predict cell types in a Seurat object with Seurat label transfer}
\usage{
seurat_predict(
  seu.q,
  seu.r,
  ref_id = "labels",
  assay.q = "RNA",
  assay.r = "RNA",
  down.sample = 500,
  pred_label = "seu.pred",
  kf = NA,
  ka = 5
)
}
\description{
Predict cell types in a Seurat object using a reference Seurat object.
An annotated Seurat object is required. The list of marker or features is required.
Predictions of cell type for each cell are made using anchors. To increase speed use down.sample
to down sample the reference by the cell types. Markers is a list of the flow cytometery antibodies,
a vector of selected features(genes) is possible. The predictions will be added to the
Seurat object as prediction name.
}
