% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{flowset_to_csv}
\alias{flowset_to_csv}
\title{Create a dataframe from a flowset object and option to save csv}
\usage{
flowset_to_csv(flowset, output_path, save.csv = FALSE)
}
\description{
Takes in a flowset object and creates a dataframe. By default no output path is added and a data frame is returned.
The name of each sample will be added as the column "Sample"
into the final dataframe. To save the dataframe directly as a csv set save.csv = filepath/tosave/
The filename will be the name of the flowset object input into the function.
}
\examples{
flowset_to_csv(flowset)
flowset_to_csv(flowset, output_path = "path/to/location/", save.csv = TRUE)
}
