% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{get_annotation}
\alias{get_annotation}
\title{Predict the cell type annotation for each cluster}
\usage{
get_annotation(
  seu,
  seu.cluster,
  seu.label,
  top_n = 3,
  filter_out = "none",
  Label = "Label"
)
}
\description{
This function creates a data frame with the predicted cell type for each cluster. The
Seurat object with clusters and predicted labels must be input. The cluster meta data slot
to label and the predicted labels slot must both be input. Sometime many cells are predicted
as unknown or another unwanted type. The option filter_out = c("unknown","mixed") can be used to
ignore these cell types if they are the most frequent cell type in a given cluster and annotate
using the second most frequent cell type.
seu.cluster = seu$RNA_snn_res.0.5
seu.label = seu$seu.pred or seu$rfm.pred
top_n = 2 (this is the number of cell types to visualize)
Only the most frequent cell type for each cluster is used for the cluster annotation.
}
