% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{harmonize}
\alias{harmonize}
\title{Transform and align samples in flowset object}
\usage{
harmonize(
  flowset,
  processing = "retro",
  two_peaks = c(9:length(colnames(transformed_flowset))),
  one_peak = c(1:8),
  threshold = 0.01
)
}
\description{
This function to take in a folder/file path of fsc flow cytometer file selecting the FSC-A
value for each channel.  Each sample file in the folder will be a slot in the flowset object.
The samples will be names by the file names and can be renamed. The channel names will automatically
be added from within the flow data. If only a subset of cells is desired downsampling is an option.
}
