% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{plot_corr}
\alias{plot_corr}
\title{Plot correlation assignment model (CAM) results}
\usage{
plot_corr(df, threshold = 0, min_cells = 100)
}
\description{
The function requires the data frame output from the find_correlation function
}
