% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{plotdensity_flowset}
\alias{plotdensity_flowset}
\title{Create a density plot from a flowset object}
\usage{
plotdensity_flowset(flowset)
}
\description{
Takes in a flowset object. Plot density plots for each channel in multichannel fsc
Flow Cytometry data. Creates stacked plots where each row is a sample in the
set object.
}
