% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{plotmean}
\alias{plotmean}
\title{Plot mean values by group in dotplot or heatmap}
\usage{
plotmean(
  plot_type = "heatmap",
  seu,
  group,
  markers,
  var_names,
  slot = "scale.data",
  xlab = "Cell Types",
  ylab = "Antibodies",
  var1order,
  var2order
)
}
\description{
This function takes a Seurat object and creates a heatmap or dotplot.
This function takes a Seurat object and a list of variables to plot. One
bar chart is created for each variable in the list.
var.list is a list defining seurat metadata slots
xgroup is the cell types or x axis grouping variable
varnames is a character vector of labels for the x axis of the plots
Example :plotproportions(seu.q, var.list = var.list, xgroup = seu.q$cell.types, varnames = varnames)
}
