% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{Prep_for_stats}
\alias{Prep_for_stats}
\title{Prepare a data frame for statistics}
\usage{
Prep_for_stats(seu, marker_list, variables, marker_name = "Marker")
}
\description{
This function takes a Seurat object and creates a data frame with the expression
values for each cell for each marker.  Columns are created for all the indicated
variables.  All the variables must exist as meta data in the Seurat object.
}
