% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{proportionplots}
\alias{proportionplots}
\title{Plot barcharts of proportion of cell types grouped by a list of variables}
\usage{
proportionplots(
  seu,
  seu.var,
  seu.lable,
  groups = "Sample",
  my_colours = "default"
)
}
\description{
This function takes a Seurat object and creates a boxplot each variable in the list.
}
