% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeR.R
\name{RFM_train}
\alias{RFM_train}
\title{Train a Random Forest Classifier to predict cell types in a Seurat object}
\usage{
RFM_train(
  seurate_object,
  marker_list,
  annotations = seu$CellTypes,
  split = c(0.8, 0.2),
  downsample = "none",
  seed = 222,
  mytry = c(6:12),
  maxnodes = c(12:25),
  trees = c(250, 500, 1000, 2000),
  start_node = 12,
  cores_to_use = 4
)
}
\description{
Requires a Seurat object for the labels to be predicted. AB_list is the list of markers(antibodies)
or selected features(genes) to input into the training model. An annotated Seurat object of matching
data is required. The location of the annotations metadata needs to be designated:
annotated_seurat_object$MetadataAnnotations_slot. The training data in the annotated Seurat object
needs to be split.  If the number of cells is high the function may not run.  If this is the case
use downsample to reduce the number of cells.  The starting data will be randomly selected from to
reduce the total size. This data will then be split in the defined ratio of training to test data.
A random seed can be set to repeat the training with different random starts and random splitting of
data. The tuning parameters for the random forst model can be optomized. The best values are
selected for the final model from mytry, maxnodes, trees, and start_node.
mtry: number of features to draw at once
maxnodes: the terminal number of nodes trees in the forest can have.
trees: the number of trees to be made.
start_node: number of starting nodes in the trees.
}
